# docs/OpenAPI.yaml
openapi: '3.0.3'
info:
  title: Ilhop Manager - OpenAPI 3.0
  description: |- 
    This is a Ilhop Manager Server based on the OpenAPI 3.0 specification.
  version: '1.0'
servers:
  - url: https://api.server.test/v1
tags:
  - name: table
  - name: customer
  - name: menu
  - name: order
  - name: order item
  - name: external
paths:
  /example:
      post: # get/post/put/delete 중 하나
        tags:
          - example
        summary: 예시 endpoint입니다.
        description: Example이라는 object를 불러옵니다.
        operationId: getExample
        parameters:
          - name: path_param
            in: path
            description: 예시 description입니다.
            required: true
            schema:
              type: string
          - name: query_param
            in: query
            description: 예시 description입니다.
            required: false
            explode: true
            schema:
              type: string
              default: option_a
              enum:
                - option_a
                - option_b
                - option_c
          - name: query_param2
            in: query
            description: 예시 description입니다.
            required: true
            explode: true
            schema:
              type: array
              items:
                type: string
        requestBody:
          description: 예시 description입니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/Error"
          '400':
            description: Bad request
          '404':
            description: Not found in the system
          default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"    
  /tables:
    get:
      tags:
        - table
      summary: 전체 테이블 및 현재 이용 현황 조회
      description: |
        모든 테이블의 위치(grid) 정보와 이용 가능 여부를 반환합니다. 
        만약 테이블에 활성화된(isActive=true) 손님이 있다면 해당 손님의 정보(입장 시각 등)를 포함합니다.
      operationId: getTables
      responses:
        '200':
          description: 성공적으로 테이블 목록을 불러왔습니다.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableStatus'
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - table
      summary: "새로운 테이블 추가 또는 배치 설정"
      description: |
        매장에 새로운 테이블을 추가하거나 그리드 상의 위치를 지정합니다. 
        백엔드에서는 `table_id`를 UUID로 자동 생성하며, 
        동일한 `gridRow`, `gridCol` 위치에 중복된 테이블이 오지 않도록 검증합니다.
      operationId: createTable
      requestBody:
        description: "추가할 테이블의 배치 정보"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreateRequest'
      responses:
        '201':
          description: "테이블 생성 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableStatus'
        '400':
          description: "잘못된 요청 (데이터 형식 오류 등)"
        '409':
          description: "충돌 발생 (이미 해당 위치에 테이블이 존재하거나 테이블 번호 중복)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /tables/{table_id}:
    patch:
      tags:
        - table
      summary: "테이블 정보 수정"
      description: |
        특정 테이블의 번호(`tableNum`)를 변경하거나, 
        테이블 파손 등의 사유로 이용 가능 여부(`isAvailable`)를 업데이트할 때 사용합니다.
      operationId: updateTable
      parameters:
        - name: tableId
          in: path
          description: "수정할 테이블의 UUID"
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: "수정할 테이블 정보"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdateRequest'
      responses:
        '200':
          description: "테이블 정보 수정 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableStatus'
        '400':
          description: "잘못된 요청 데이터"
        '404':
          description: "해당 테이블을 찾을 수 없음"
        '409':
          description: "테이블 번호 중복 등 데이터 충돌"
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - table
      summary: "특정 테이블 삭제"
      description: |
        지정한 UUID를 가진 테이블을 시스템에서 삭제합니다. 
        단, 해당 테이블에 아직 `isActive: true`인 손님이 있거나 결제되지 않은 주문이 연결되어 있을 경우 
        데이터 무결성을 위해 삭제가 제한될 수 있습니다.
      operationId: deleteTable
      parameters:
        - name: tableId
          in: path
          description: "삭제할 테이블의 UUID"
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "삭제 성공 (반환할 콘텐츠 없음)"
        '400':
          description: "잘못된 요청 (활성 고객이 있는 테이블 등)"
        '404':
          description: "해당 테이블을 찾을 수 없음"
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /customers:
    get:
      tags:
        - customer
      summary: "조건에 맞는 고객 목록 조회"
      description: |
        테이블 번호나 활성화 여부 등의 조건을 필터로 사용하여 고객 정보를 조회합니다.
        주문 웹페이지 접속 시 `tableNum={n}&isActive=true` 파라미터를 사용하여 
        현재 해당 테이블을 이용 중인 고객 1명의 정보를 가져오는 데 주로 사용됩니다.
      operationId: getCustomers
      parameters:
        - name: tableNum
          in: query
          description: "조회할 테이블 번호"
          required: false
          schema:
            type: integer
            example: 7
        - name: isActive
          in: query
          description: "활성화(이용 중) 여부 필터"
          required: false
          schema:
            type: boolean
            default: true
            example: true
      responses:
        '200':
          description: "조회 성공 (조건에 맞는 고객이 없을 경우 빈 배열 `[]` 반환)"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerBrief'
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - customer
      summary: 새 손님 입장 (테이블 점유)
      description: |
        특정 테이블에 새로운 손님 세션을 생성합니다. 
        백엔드에서 해당 테이블의 `table_id`를 조회하고, 현재 시각을 `entry_time`으로 설정하여 
        `isActive: true` 상태의 고객 레코드를 생성합니다.
      operationId: createCustomer
      requestBody:
        description: 손님이 앉은 테이블 번호
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateRequest'
      responses:
        '201':
          description: 손님 입장 처리 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerBrief'
        '400':
          description: 잘못된 요청 (존재하지 않는 테이블 번호 등)
        '409':
          description: 해당 테이블이 이미 사용 중임
        default:
          description: 예기치 못한 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /customers/{customer_id}/order-summary:
    get:
      tags:
        - customer
      summary: "특정 고객의 결제 완료 주문 내역 조회"
      description: |
        특정 고객 ID를 필터로 사용하여 해당 고객이 주문한 내역 중 
        결제 승인(`isPaid=true`)이 완료된 항목들만 가져옵니다.
        동일한 메뉴에 대한 주문 내역은 수량을 합산하여 하나로 묶어서 가져옵니다.
        주문 페이지의 '주문 내역' 탭에서 사용됩니다.
      operationId: getCustomerOrderSummary
      parameters:
        - name: customerId
          in: path
          description: "조회할 고객의 UUID"
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440005"
        - name: isPaid
          in: query
          description: "결제 완료 여부 필터 (기본값 true)"
          required: false
          schema:
            type: boolean
            default: true
            example: true
      responses:
        '200':
          description: "성공적으로 주문 내역을 불러옴"
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/OrderSummaryResponse'
        '400':
          description: "잘못된 파라미터 형식"
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /menus:
    get:
      tags:
        - menu
      summary: 전체 메뉴 목록 조회
      description: |
        주문 페이지 접속 시 노출할 전체 메뉴 목록을 가져옵니다. 
        각 메뉴의 가격, 이미지 URL, 선택 가능한 옵션 리스트를 포함합니다.
      operationId: getMenuList
      responses:
        '200':
          description: 메뉴 목록을 성공적으로 불러왔습니다.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuResponse'
        '500':
          description: 서버 내부 오류 (메뉴 데이터를 불러올 수 없음)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - menu
      summary: 새로운 메뉴 추가
      description: |
        메뉴판에 새로운 메뉴를 추가합니다.
        백엔드에서는 `menu_id`를 UUID로 자동 생성합니다.
      operationId: createMenu
      requestBody:
        description: 추가할 메뉴의 정보
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuCreateRequest'
      responses:
        '201':
          description: 메뉴 추가 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
        '400':
          description: 잘못된 요청 데이터
        default:
          description: 예기치 못한 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders:
    get:
      tags:
        - order
      summary: "전체 주문 내역 리스트 조회 (관리자용)"
      description: |
        시스템의 모든 주문 내역을 개별 건별로 조회합니다. 
        캐셔가 입금자명과 금액을 대조하여 `isPaid`를 업데이트하기 위한 용도로 사용됩니다.
      operationId: getOrders
      parameters:
        - name: isPaid
          in: query
          description: "미입금 내역만 보려면 false 필터 사용"
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: "전체 주문 리스트 반환"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDetail'
    
    post:
      tags:
        - order
      summary: "새 주문 생성 (장바구니 결제)"
      description: |
        localStorage의 장바구니 데이터를 받아 주문을 생성합니다.
        백엔드는 `orders` 테이블에 기본 정보를 저장하고, 상세 메뉴들을 `order_items`에 기록합니다.
        장바구니 안에서 같은 `order_id`와 `selected_option`을 가진 항목은 반드시 하나의 `order_item`으로 묶여 있어야 합니다.
        즉, 하나의 (`order_id`, `menu_id`, `selected_option`) 조합으로 유일한 `order_item`을 특정할 수 있어야 합니다.
      operationId: createOrder
      requestBody:
        description: 주문 정보 및 장바구니에 담긴 메뉴 리스트
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: 주문 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreateResponse'
        '400':
          description: 잘못된 요청 데이터
        '404':
          description: 고객 정보 또는 메뉴 정보를 찾을 수 없음
        default:
          description: 예기치 못한 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /orders/{order_id}/payment:
    patch:
      tags:
        - order
      summary: "주문 결제 상태 수동 업데이트"
      description: |
        현금 결제 완료 또는 계좌 이체 수동 확인 시 사용합니다. 
        해당 주문(`order`)의 `isPaid`를 `true`로 변경합니다.
      operationId: updateOrderisPaid
      parameters:
        - name: orderId
          in: path
          description: "결제 처리할 주문의 UUID"
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: "결제 상태 변경 정보"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPaymentUpdateRequest'
      responses:
        '200':
          description: "결제 상태 업데이트 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderBrief'
        '404':
          description: "존재하지 않는 주문 ID"
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /orders/{order_id}/memo:
    patch:
      tags:
        - order
      summary: "주문 비고(메모) 업데이트"
      description: |
        특정 주문에 대해 관리자(캐셔)가 메모를 작성하거나 수정합니다. 
        빈 문자열을 보낼 경우 메모가 삭제(또는 초기화)됩니다.
      operationId: updateOrderMemo
      parameters:
        - name: orderId
          in: path
          description: "메모를 추가할 주문의 UUID"
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: "업데이트할 메모 내용"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderMemoUpdateRequest'
      responses:
        '200':
          description: "메모 업데이트 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderBrief'
        '404':
          description: "해당 주문을 찾을 수 없음"
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /order-items/{order_id}/{menu_id}:
    patch:
      tags:
        - order item
      summary: "특정 주문 항목의 서빙 상태 업데이트"
      description: |
        홀서빙 담당자가 특정 메뉴의 서빙을 완료했을 때 호출합니다. 
        해당 주문 항목(`order_item`)의 `is_served` 상태를 `true`로 변경하여 
        캐셔와 홀서빙 화면에 실시간으로 반영되도록 합니다.
      operationId: updateOrderItemServedStatus
      parameters:
        - name: orderId
          in: path
          description: "대상 주문의 UUID"
          required: true
          schema:
            type: string
            format: uuid
        - name: menuId
          in: path
          description: "대상 메뉴의 UUID"
          required: true
          schema:
            type: string
            format: uuid
        - name: selectedOption
          in: query
          description: "대상 메뉴의 선택된 옵션"
          required: false
          schema:
            type: string
            example: "살구맛"
      requestBody:
        description: "변경할 서빙 상태 값"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemServedUpdateRequest'
      responses:
        '200':
          description: "서빙 상태 업데이트 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemBrief'
        '400':
          description: "잘못된 요청 파라미터"
        '404':
          description: "해당 주문 항목을 찾을 수 없음"
        default:
          description: "서버 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /external/bank-webhook:
    post:
      tags:
        - external
      summary: MacroDroid이 입급 확인을 알릴 때 사용함
      description: |
        총무의 핸드폰으로 토스 입금 알림이 오면,
        핸드폰에 설치된 MacroDroid 앱이 아래와 같은 형식의 request body를 담은 request를 백엔드로 보냄 
      operationId: confirmPayment
      requestBody:
        description: 입금자명, 송금 액수, 알림을 보낸 앱(토스) 정보를 담음
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentConfirmInfo'
      responses:
        '200':
          description: 성공적으로 입금을 확인했습니다.
        default:
            description: 서버 오류
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"

components:
  schemas:
    TableCreateRequest:
      type: object
      properties:
        tableNum:
          type: integer
          description: "손님에게 보일 테이블 번호"
          example: 15
        gridRow:
          type: integer
          description: "매장 그리드 상의 가로 행 위치"
          example: 3
        gridCol:
          type: integer
          description: "매장 그리드 상의 세로 열 위치"
          example: 5
        isAvailable:
          type: boolean
          description: "즉시 사용 가능 여부 (기본값 true)"
          default: true
          example: true
      required:
        - tableNum
        - gridRow
        - gridCol
    
    TableUpdateRequest:
      type: object
      properties:
        tableNum:
          type: integer
          description: "변경할 새로운 테이블 번호"
          example: 12
        isAvailable:
          type: boolean
          description: "테이블 사용 가능 여부 상태"
          example: false
      # PATCH이므로 모든 필드를 필수로 둘 필요는 없으나, 
      # 요청 시 최소한 하나는 보내도록 백엔드에서 검증하는 것이 좋습니다.

    TableStatus:
      type: object
      description: Table 정보에 현재 이용 중인 Customer 정보를 포함한 확장 스키마
      properties:
        tableId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        tableNum:
          type: integer
          description: "손님에게 보이는 테이블 번호"
          example: 7
        gridRow:
          type: integer
          description: "그리드 상의 세로 위치"
          example: 2
        gridCol:
          type: integer
          description: "그리드 상의 가로 위치"
          example: 3
        isAvailable:
          type: boolean
          description: "테이블 사용 가능 여부 (한 손님이 두 테이블 사용 등)"
          default: true
          example: true
        currentCustomer:
          $ref: '#/components/schemas/CustomerBrief'
      required:
        - tableId
        - tableNum
        - gridRow
        - gridCol
        - isAvailable

    CustomerCreateRequest:
      type: object
      properties:
        tableNum:
          type: integer
          description: "손님이 착석한 테이블 번호"
          example: 7
      required:
        - tableNum

    CustomerBrief:
      type: object
      nullable: true
      description: "현재 테이블을 이용 중인 고객 정보 (비어있으면 null)"
      properties:
        customerId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        entryTime:
          type: string
          format: date-time
          description: "입장 시각 (프론트엔드에서 경과 시간 계산용)"
          example: "2024-04-12T18:25:30Z"
        isActive:
          type: boolean
          example: true
      required:
        - customerId
        - entryTime 
        - isActive

    MenuCreateRequest:
      type: object
      properties:
        menuName:
          type: string
          description: 음식 또는 음료 이름
          example: "치킨 가라아게"
        price:
          type: integer
          format: int64
          description: "판매 가격 (원 단위)"
        imageUrl:
          type: string
          format: uri
          description: "메뉴 사진 경로"
          example: "https://example.com/images/chicken.jpg"
        options:
          type: array
          description: "선택 가능한 옵션 (맛, 사이즈 등)"
          items:
            type: string
          example: ["살구맛", "청포도맛", "레몬맛"]
      required:
        - menuName
        - price

    MenuResponse:
      type: object
      properties:
        menuId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        menuName:
          type: string
          description: 음식 또는 음료 이름
          example: "치킨 가라아게"
        price:
          type: integer
          format: int64
          description: "판매 가격 (원 단위)"
          example: 15000
        imageUrl:
          type: string
          format: uri
          description: "메뉴 사진 경로"
          example: "https://example.com/images/chicken.jpg"
        options:
          type: array
          description: "선택 가능한 옵션 (맛, 사이즈 등)"
          items:
            type: string
          example: ["살구맛", "청포도맛", "레몬맛"]
      required:
        - menuId
        - menuName
        - price

    OrderCreateRequest:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440005"
        totalPrice:
          type: integer
          format: int64
          example: 25000
        depositor:
          type: string
          description: "입금자명 (현금 입금 확인용)"
          example: "홍길동"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
      required:
        - customerId
        - totalPrice
        - items

    OrderCreateResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        orderTime:
          type: string
          format: date-time
          example: "2024-04-12T19:00:00Z"
        totalPrice:
          type: integer
          format: int64
          example: 30000
        depositor: # 입금자명 대조를 위해 필요함
          type: string
          example: "홍길동"
        items: # 해당 주문 건에 포함된 메뉴들
          type: array
          items:
            $ref: '#/components/schemas/OrderItemBrief'
      required:
        - orderId
        - orderTime
        - totalPrice

    OrderPaymentUpdateRequest:
      type: object
      properties:
        isPaid:
          type: boolean
          description: "변경할 결제 상태 (수동 승인 시 true)"
          example: true
      required:
        - isPaid

    OrderMemoUpdateRequest:
      type: object
      properties:
        memo:
          type: string
          maxLength: 255
          nullable: true
          description: "작성할 메모 내용 (최대 255자)"
          example: "현금으로 50,000원 지불 확인됨"
      required:
        - memo

    OrderBrief:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        isPaid:
          type: boolean
          default: false
          example: false
        memo:
          type: string
          nullable: true
          example: "현금 결제 완료"

    OrderDetail:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        tableNum: # 캐셔가 어느 테이블인지 바로 알아야 하므로 추가
          type: integer
          example: 7
        customerId: # 같은 손님이 주문한 내역만 묶어볼 수 있게하기 위해서 추가
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        orderTime:
          type: string
          format: date-time
          example: "2024-04-12T19:00:00Z"
        totalPrice:
          type: integer
          format: int64
          example: 30000
        depositor: # 입금자명 대조를 위해 필요함
          type: string
          example: "홍길동"
        isPaid:
          type: boolean
          default: false
          example: false
        memo:
          type: string
          nullable: true
          example: "현금 결제 완료"
        items: # 해당 주문 건에 포함된 메뉴들
          type: array
          items:
            $ref: '#/components/schemas/OrderItemBrief'
      required:
        - orderId
        - tableNum
        - customerId
        - orderTime
        - totalPrice
        - isPaid

    OrderSummaryResponse:
      type: object
      properties:
        totalAmount:
          type: integer
          format: int64
          description: "해당 테이블의 전체 결제 금액 합계"
          example: 25000
        orderItems:
          type: object
          $ref: "#/components/schemas/OrderItemSummaryResponse"
      required:
        - totalAmount
        - orderItems

    OrderItemRequest:
      type: object
      properties:
        menuId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        quantity:
          type: integer
          minimum: 1
          example: 2
        priceAtOrder:
          type: integer
          format: int64
          description: "주문 시점의 메뉴 가격"
          example: 8000
        selectedOption:
          type: string
          nullable: true
          example: "살구맛"
      required:
        - menuId
        - quantity
        - priceAtOrder
    
    OrderItemServedUpdateRequest:
      type: object
      properties:
        isServed:
          type: boolean
          description: "서빙 완료 여부"
          example: true
      required:
        - isServed

    OrderItemBrief:
      type: object
      properties:
        menuName:
          type: string
          example: "치킨 가라아게"
        quantity:
          type: integer
          minimum: 1
          example: 2
        selectedOption:
          type: string
          nullable: true
          example: "살구맛"
        isServed:
          type: boolean
          default: false
          example: false
      required:
        - menuName
        - quantity
        - isServed

    OrderItemSummaryResponse:
      type: object
      properties:
        menuName:
          type: string
          example: "좋은토닉"
        totalQuantity:
          type: integer
          description: "메뉴+옵션 조합의 합계 수량"
          example: 2
        unitPrice:
          type: integer
          format: int64
          example: 5000
        selectedOption:
          type: string
          nullable: true
          description: "선택된 옵션명 (없으면 null)"
          example: "살구맛"
      required:
        - menuName
        - totalQuantity
        - unitPrice
    
    PaymentConfirmInfo:
      type: object
      properties:
        title:
          description: 알림 제목
          type: string
        message:
          description: 알림 본문
          type: string
        app:
          description: 알림 보낸 앱
          type: string

    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message